added tests for testing full threadpool and queue and manual canceling
added abort policy to threadpool configuration. added unit-test for testing this on a full threadpool with a full queue.
added unit-test for manually canceling execution of code.
sunsetexecutor processes now have an automated timeout.
added new class for public messages.
other code improvements.

+++ src/main/java/com/syssec/sunsetmiddleware/executor/SunsetExecutor.java
@@ -7,6 +7,10 @@ import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
 import java.time.Duration;
 import java.time.Instant;
+import java.util.concurrent.TimeoutException;
+
+import com.syssec.sunsetmiddleware.messages.SunsetGlobalMessages;
+import com.syssec.sunsetmiddleware.threadpool.SunsetThreadPoolConfiguration;
 
 /**
  * Class for managing the execution of the received code with sunset (via
@@ -16,9 +20,19 @@ import java.time.Instant;
  *
  */
 public class SunsetExecutor {
-
 	private String sunsetPath = "sunset.jar";
 	private Process process;
+	private int timeoutSeconds;
+
+	public SunsetExecutor() {
+		this.timeoutSeconds = SunsetThreadPoolConfiguration.KEEP_ALIVE_SECONDS_DEFAULT + 10;
+		try {
+			this.process = Runtime.getRuntime().exec("java -jar " + sunsetPath + " --cmd");
+		} catch (IOException e) {
+			this.destroyProcess();
+			e.printStackTrace();
+		}
+	}
 
 	/**
 	 * Method used for executing sunset via commandline using the code sent by the
@@ -26,10 +40,12 @@ public class SunsetExecutor {
 	 * 
 	 * @param code - Code received by the user
 	 * @return result of code execution as plain text (String)
+	 * @throws TimeoutException
+	 * @throws InterruptedException
 	 */
-	public String executeCommand(String receivedCode) {
+	public String executeCommand(String receivedCode) throws TimeoutException, InterruptedException {
 		if (receivedCode.isEmpty()) {
-			throw new IllegalArgumentException("Empty input code received!");
+			throw new IllegalArgumentException(SunsetGlobalMessages.EMPTY_CODE_RECEIVED);
 		}
 
 		String code = receivedCode;
@@ -38,12 +54,13 @@ public class SunsetExecutor {
 		if (!code.endsWith("END"))
 			code = code + "\nEND\n";
 
-		System.out.println("[INFO: Executing sunset in commandline with provided code!]");
-
-		Instant startTime = Instant.now();
+		System.out.println("[INFO:] " + SunsetGlobalMessages.SUNSET_EXECUTION_VIA_COMMANDLINE);
 
 		try {
-			this.process = Runtime.getRuntime().exec("java -jar " + sunsetPath + " --cmd");
+			Instant startTime = Instant.now();
+
+			long timeoutMilliseconds = this.timeoutSeconds * 1000;
+			long timoutTime = startTime.toEpochMilli() + timeoutMilliseconds;
 
 			BufferedReader in = new BufferedReader(new InputStreamReader(process.getInputStream()));
 			BufferedWriter out = new BufferedWriter(new OutputStreamWriter(process.getOutputStream()));
@@ -51,27 +68,37 @@ public class SunsetExecutor {
 			out.write(code);
 			out.flush();
 
+			while (this.isAlive(process)) {
+				if (System.currentTimeMillis() > timoutTime) {
+					this.destroyProcess();
+					throw new TimeoutException("Timeout occurred after " + this.timeoutSeconds + " seconds!");
+				}
+			}
+
 			in.readLine();
 			result = "";
 			String line = null;
 			while ((line = in.readLine()) != null) {
 				result += line + "\n";
 			}
-
-			this.destroyProcess();
-
 			Instant endTime = Instant.now();
-
 			long elapsedTime = Duration.between(startTime, endTime).toMillis();
-
 			System.out.println("[INFO: Duration of sunset execution: " + elapsedTime + "ms]");
 
 			return result.trim();
+
 		} catch (IOException e) {
 			this.destroyProcess();
-			System.out.println("[ERROR: There was an exception when trying to execute sunset!]");
-			System.out.println(e.getMessage());
 			return e.getMessage();
+		}
+	}
+
+	private boolean isAlive(Process p) {
+		try {
+			p.exitValue();
+			return false;
+		} catch (IllegalThreadStateException e) {
+			return true;
 		}
 	}
 
@@ -81,6 +108,14 @@ public class SunsetExecutor {
 
 	public boolean isProcessAlive() {
 		return this.process.isAlive();
+	}
+
+	public void setTimeoutSeconds(int timeoutSeconds) {
+		if (timeoutSeconds <= 0) {
+			throw new IllegalArgumentException("Value for TimeoutSeconds must be >0!");
+		}
+
+		this.timeoutSeconds = timeoutSeconds;
 	}
 
 }
